deleting post

+++ graphql-course/graphql-prisma/src/resolvers/Mutation.js
@@ -113,7 +113,18 @@ const Mutation = {
       info
     );
   },
-  async deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!postExists) throw new Error(`Unable to delete post`);
+
     return await prisma.mutation.deletePost({ where: { id: args.id } }, info);
   },
   async createComment(parent, args, { prisma }, info) {
